services: Refactor into directories
