+++ src/services/fs/backend.rs
@@ -12,31 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Posix file system support.
-
 use std::collections::HashMap;
 use std::fs;
 use std::io::SeekFrom;
 use std::path::PathBuf;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::Context;
-use std::task::Poll;
 
 use anyhow::anyhow;
 use async_trait::async_trait;
 use blocking::unblock;
 use blocking::Unblock;
 use futures::io;
-use futures::ready;
 use futures::AsyncReadExt;
 use futures::AsyncSeekExt;
 use futures::AsyncWriteExt;
-use log::debug;
 use log::error;
 use log::info;
 use metrics::increment_counter;
 
+use super::error::parse_io_error;
+use super::object_stream::Readdir;
 use crate::error::Error;
 use crate::error::Kind;
 use crate::error::Result;
@@ -50,7 +45,6 @@ use crate::ops::OpStat;
 use crate::ops::OpWrite;
 use crate::Accessor;
 use crate::BoxedAsyncReader;
-use crate::Object;
 
 #[derive(Default, Debug)]
 pub struct Builder {
@@ -320,113 +314,8 @@ impl Accessor for Backend {
             e
         })?;
 
-        let rd = Readdir {
-            acc: Arc::new(self.clone()),
-            root: self.root.clone(),
-            path: args.path.clone(),
-            rd: Unblock::new(f),
-        };
+        let rd = Readdir::new(Arc::new(self.clone()), &self.root, &args.path, f);
 
         Ok(Box::new(rd))
-    }
-}
-
-struct Readdir {
-    acc: Arc<dyn Accessor>,
-    root: String,
-    path: String,
-
-    rd: Unblock<std::fs::ReadDir>,
-}
-
-impl futures::Stream for Readdir {
-    type Item = Result<Object>;
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        match ready!(Pin::new(&mut self.rd).poll_next(cx)) {
-            None => {
-                debug!("object {} list done", &self.path);
-                Poll::Ready(None)
-            }
-            Some(Err(e)) => {
-                error!("object {} stream poll_next: {:?}", &self.path, e);
-                Poll::Ready(Some(Err(parse_io_error(e, "list", &self.path))))
-            }
-            Some(Ok(de)) => {
-                // NOTE: metadata is syscall.
-                let de_meta = de.metadata().map_err(|e| {
-                    let e = parse_io_error(e, "list", &de.path().to_string_lossy());
-                    error!("object {:?} metadata: {:?}", &de.path(), e);
-                    e
-                });
-                if let Err(e) = de_meta {
-                    return Poll::Ready(Some(Err(e)));
-                }
-                let de_meta = de_meta.unwrap();
-
-                let de_path = de.path();
-                let de_path = de_path.strip_prefix(&self.root).map_err(|e| {
-                    let e = Error::Object {
-                        kind: Kind::Unexpected,
-                        op: "list",
-                        path: de.path().to_string_lossy().to_string(),
-                        source: anyhow::Error::from(e),
-                    };
-                    error!("object {:?} path strip_prefix: {:?}", &de.path(), e);
-                    e
-                })?;
-                let path = de_path.to_string_lossy();
-
-                let mut o = Object::new(self.acc.clone(), &path);
-
-                let meta = o.metadata_mut();
-                meta.set_complete();
-                if de_meta.is_dir() {
-                    meta.set_mode(ObjectMode::DIR);
-                } else {
-                    meta.set_mode(ObjectMode::FILE);
-                }
-                meta.set_content_length(de_meta.len());
-                meta.set_complete();
-
-                debug!(
-                    "object {} got entry, path: {}, mode: {}",
-                    &self.path,
-                    meta.path(),
-                    meta.mode()
-                );
-                Poll::Ready(Some(Ok(o)))
-            }
-        }
-    }
-}
-
-/// Parse all path related errors.
-///
-/// ## Notes
-///
-/// Skip utf-8 check to allow invalid path input.
-fn parse_io_error(err: std::io::Error, op: &'static str, path: &str) -> Error {
-    use std::io::ErrorKind;
-
-    match err.kind() {
-        ErrorKind::NotFound => Error::Object {
-            kind: Kind::ObjectNotExist,
-            op,
-            path: path.to_string(),
-            source: anyhow::Error::from(err),
-        },
-        ErrorKind::PermissionDenied => Error::Object {
-            kind: Kind::ObjectPermissionDenied,
-            op,
-            path: path.to_string(),
-            source: anyhow::Error::from(err),
-        },
-        _ => Error::Object {
-            kind: Kind::Unexpected,
-            op,
-            path: path.to_string(),
-            source: anyhow::Error::from(err),
-        },
     }
 }
