+++ src/services/memory/backend.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! In memory backend support.
-
 use std::collections::HashMap;
 use std::pin::Pin;
 use std::sync::Arc;
